Don't use SQL for sampling without replacement. Instead, we sample with fraction 1.0 and then take the 1st numSamples

+++ core/src/main/java/io/ddf/facades/ViewsFacade.java
@@ -43,7 +43,7 @@ public class ViewsFacade implements IHandleViews {
   }
 
   @Override
-  public DDF getRandomSampleByNum(long numSamples, boolean withReplacement,
+  public DDF getRandomSampleByNum(int numSamples, boolean withReplacement,
                                   int seed) {
     return mViewHandler.getRandomSampleByNum(numSamples, withReplacement, seed);
   }
